Fix signer new token
