+++ electron/messages/index.js
@@ -14,9 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
-const { spawn } = require('child_process');
-
 const runParity = require('./runParity');
+const signerNewToken = require('./signerNewToken');
 
 /**
  * Handle all asynchronous messages from renderer to main.
@@ -28,26 +27,9 @@ module.exports = (event, arg) => {
       break;
     }
     case 'signer-new-token': {
-      // Generate a new token if we can find the parity binary
-      if (!global.parityInstallLocation) { return; }
-      const paritySigner = spawn(global.parityInstallLocation, ['signer', 'new-token']);
-
-      // Listen to the output of the previous command
-      paritySigner.stdout.on('data', (data) => {
-        // If the output line is xxxx-xxxx-xxxx-xxxx, then it's our token
-        const match = data.toString().match(/[a-zA-Z0-9]{4}(-)?[a-zA-Z0-9]{4}(-)?[a-zA-Z0-9]{4}(-)?[a-zA-Z0-9]{4}/);
-
-        if (match) {
-          const token = match[0];
-
-          // Send back the token to the renderer process
-          event.sender.send('asynchronous-reply', token);
-          paritySigner.kill(); // We don't need the signer anymore
-        }
-      });
+      signerNewToken(event);
       break;
     }
     default:
   }
-}
-  ;
+};
